Move FilterableListMixin into models
This change moves FilterableListMixin and CategoryFilterableMixin into the v1/models directory.
