+++ cfgov/v1/models/browse_filterable_page.py
@@ -11,11 +11,10 @@ from wagtail.search import index
 from v1 import blocks as v1_blocks
 from v1.atomic_elements import molecules, organisms
 from v1.models.base import CFGOVPage
-from v1.models.learn_page import EnforcementActionPage, EventPage
-from v1.util.filterable_list import (
-    FilterableListMixin,
-    CategoryFilterableMixin,
+from v1.models.filterable_list_mixins import (
+    CategoryFilterableMixin, FilterableListMixin
 )
+from v1.models.learn_page import EnforcementActionPage, EventPage
 
 
 class BrowseFilterableContent(StreamBlock):
