+++ cfgov/v1/models/sublanding_filterable_page.py
@@ -9,9 +9,8 @@ from wagtail.search import index
 from v1 import blocks as v1_blocks
 from v1.atomic_elements import molecules, organisms
 from v1.models.base import CFGOVPage
-from v1.util.filterable_list import (
-    FilterableListMixin,
-    CategoryFilterableMixin,
+from v1.models.filterable_list_mixins import (
+    CategoryFilterableMixin, FilterableListMixin
 )
